Move landing page component directly into route component
